update build to run augurv6 auspicev2
When I initially made the build, augur v6 had not yet been released. This (along with Auspice v2) is a big version change. Therefore, to ensure that this build would remain stable for longer, I wanted to transition it over to Augur v6 which makes Auspice v2 JSONs. This required changes to the Snakefile as well as the auspice config file.