fix flimsiness of export
Originally when I exported the dataframes as CSV files I had reindexed them and relabeled the index column. But this meant that if you went back to run the earlier code, it didn't work because the index had been reset. To improve this, I stopped the reindexing process (since it was just for re-labeling) and instead used the pandas dataframe export flag for renaming columns.